pyproject.toml
[tool.poetry]
name = "bigbuild"
version = "0.1.0"
description = "🛠️ BigBuildBench: Benchmarking the repo-level understanding capability of your LLMs by reconstructing project build-file."
authors = ["B3 Group <b3@BigBuildBench.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "bigbuild"}]

[tool.poetry.dependencies]
python = "^3.10"
fire = "^0.6.0"
pygithub = "^2.4.0"
alive-progress = "^3.1.5"
tabulate = "^0.9.0"
toml = "^0.10.2"
poetry = "^1.8.3"
tree-sitter = "0.21.3"
tree_sitter_languages = ">=1.10.2"
datasets = "^3.0.1"
lxml = "^5.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pre-commit = "^3.8.0"

[tool.poetry.group.eval.dependencies]
docker = "^7.1.0"


[tool.poetry.group.llm.dependencies]
anthropic = "^0.34.2"
openai = "^1.50.0"
transformers = "^4.45.1"
google-generativeai = "^0.8.2"
stop-sequencer = "^1.2.3"
vllm = "^0.6.2"

[tool.poetry.group.pattern.dependencies]
tree-sitter-c-sharp = "^0.23.1"
tree-sitter-rust = "^0.23.2"
tree-sitter-python = "^0.23.4"
tree-sitter-typescript = "^0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
